+++ mavve/src/pages/MyPage/MyPage.jsx
@@ -1,17 +1,15 @@
 import React, { useState, useRef } from "react";
 import { useNavigate } from "react-router-dom";
 import * as S from "./MyPage.style";
-import WaveImg from "../../assets/MyPage/wave.svg";
-import DefaultProfile from "../../assets/MyPage/defaultProfile.svg";
-import EditIcon from "../../assets/MyPage/editIcon.svg";
-import EditIconHover from "../../assets/MyPage/editIconHover.svg";
 import ProfileTestImg from "../../assets/MyPage/profileTest.png";
-import XIcon from "../../assets/MyPage/xIcon.svg";
-import ImgUploadIcon from "../../assets/MyPage/imgUploadIcon.svg";
 import OneLineNote from "../../components/Common/OneLineNote";
 import RoomComponent from "../../components/Common/RoomComponent";
 import SideBar from "../../components/Common/SideBar";
 import TopBar from "../../components/Common/TopBar";
+import CreateRoomBtn from "../../assets/MyPage/createRoomBtn.svg";
+import CreateRoomBtnHover from "../../assets/MyPage/createRoomBtnHover.svg";
+import Profile from "./Profile";
+import ProfileEditModal from "./ProfileEditModal";
 
 export default function MyPage() {
   const [user, setUser] = useState({
@@ -22,7 +20,7 @@ export default function MyPage() {
   });
 
   const navigate = useNavigate();
-  const fileInputRef = useRef(null);
+
   const [isHovered, setIsHovered] = useState(false);
   const [isEditing, setIsEditing] = useState(false);
   const [nameInput, setNameInput] = useState(user.name);
@@ -164,106 +162,44 @@ export default function MyPage() {
         </S.SidebarContainer>
         <S.Main>
           <S.ViewArea>
-            <S.ProfileContainer>
-              <S.WaveImg src={WaveImg} alt="wave" />
-              <S.ProfileContent>
-                <S.ProfileImage
-                  src={user.profileImg || DefaultProfile}
-                  alt="프로필 이미지"
-                />
-                <S.ProfileTextContainer>
-                  <S.Nickname>{user.name}</S.Nickname>
-                  <S.InfoContainer>
-                    <S.UserCount>
-                      <S.UserText>
-                        내 플레이리스트 {user.playlistCount}개
-                      </S.UserText>
-                      <S.UserText>내 방 {user.roomCount}개</S.UserText>
-                    </S.UserCount>
-                    <S.EditIcon
-                      src={isHovered ? EditIconHover : EditIcon}
-                      alt="편집 아이콘"
-                      onMouseEnter={() => setIsHovered(true)}
-                      onMouseLeave={() => setIsHovered(false)}
-                      onClick={() => setIsEditing(true)}
-                    />
-                  </S.InfoContainer>
-                </S.ProfileTextContainer>
-              </S.ProfileContent>
-            </S.ProfileContainer>
-
+            <Profile user={user} onEditClick={() => setIsEditing(true)} />
             {isEditing && (
-              <S.ProfileEditBackground onClick={() => setIsEditing(false)}>
-                <S.ProfileEditBox onClick={(e) => e.stopPropagation()}>
-                  <S.ProfileEditHeader>
-                    프로필 편집
-                    <img
-                      src={XIcon}
-                      alt="닫기"
-                      onClick={() => setIsEditing(false)}
-                    />
-                  </S.ProfileEditHeader>
-                  <S.ProfileEditArea>
-                    <S.ProfileImgContainer>
-                      <S.ProfileImgEdit>
-                        <label htmlFor="profileImgInput">
-                          {user.profileImg ? (
-                            <img src={user.profileImg} alt="프로필 이미지" />
-                          ) : (
-                            <img
-                              src={ImgUploadIcon}
-                              alt="업로드 아이콘"
-                              style={{ width: "3rem", height: "3rem" }}
-                            />
-                          )}
-                        </label>
-                      </S.ProfileImgEdit>
-                      <input
-                        type="file"
-                        id="profileImgInput"
-                        accept="image/*"
-                        ref={fileInputRef}
-                        style={{ display: "none" }}
-                        onChange={handleImgUplad}
-                      />
-                    </S.ProfileImgContainer>
-                    <S.ProfileEditInputs>
-                      이름
-                      <input
-                        type="text"
-                        value={nameInput}
-                        maxLength={10}
-                        onChange={(e) => setNameInput(e.target.value)}
-                      />
-                      <S.SaveButton
-                        onClick={() => {
-                          setUser((prev) => ({ ...prev, name: nameInput }));
-                          setIsEditing(false);
-                        }}
-                      >
-                        저장하기
-                      </S.SaveButton>
-                    </S.ProfileEditInputs>
-                  </S.ProfileEditArea>
-                </S.ProfileEditBox>
-              </S.ProfileEditBackground>
+              <ProfileEditModal
+                user={user}
+                nameInput={nameInput}
+                setNameInput={setNameInput}
+                setUser={setUser}
+                onClose={() => setIsEditing(false)}
+                onImageUpload={handleImgUplad}
+              />
             )}
-
             <S.OneLineNoteContainer>
               <S.Title>오늘의 한 줄 일기</S.Title>
               <OneLineNote />
             </S.OneLineNoteContainer>
+
             <S.MyRoomArea>
-              <S.Title
-                onClick={() => {
-                  if (myRooms.length >= 8) {
-                    navigate("/mypage/myroom");
-                  }
-                }}
-                style={{ cursor: myRooms.length >= 8 ? "pointer" : "default" }}
-              >
-                내가 만든 방
-              </S.Title>
+              <S.MyRoomHeader>
+                <S.Title
+                  onClick={() => {
+                    if (myRooms.length >= 8) {
+                      navigate("/mypage/myroom");
+                    }
+                  }}
+                  style={{
+                    cursor: myRooms.length >= 8 ? "pointer" : "default",
+                  }}
+                >
+                  내가 만든 방
+                </S.Title>
+                <S.CreateRoomBtn
+                  src={isHovered ? CreateRoomBtnHover : CreateRoomBtn}
+                  alt="방 만들기"
+                  onClick={() => navigate("/room")}
+                  onMouseEnter={() => setIsHovered(true)}
+                  onMouseLeave={() => setIsHovered(false)}
+                />
+              </S.MyRoomHeader>
               <S.MyRoomContainer
                 ref={myRoomRef}
                 onWheel={handleHorizontalScroll(myRoomRef)}
